Rename DebugUtils to Development namespace
